+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - `Modifier::normalizeIp` returns the host normalized for IPv6 and IPv4 addresses
 - `Modifier::normalizeHost` returns the host as normalized by the WHATWG algorithm
 - `Modifier::with*` method from the underlying `Uri` object are proxy to improve DX.
+- `Modifier::uri` method which returns the underlying URI object used by the Modifier class.
 - `Modifier::prefixQueryPairs` and `Modifier::prefixQueryParameters` to prefix Query using the pair key or the parameter names
 - `Query::decoded` the string representation of the component decoded.
 - `URLSearchParams::decoded` the string representation of the component decoded.
@@ -24,6 +25,7 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - `HierarchicalPath::normalized`
 - `Datapath::normalized`
 - Support for `Uri\Rfc3986\Uri`
+- Support for `Uri\WhatWg\Url`
 
 ### Fixed
 
@@ -33,6 +35,7 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 ### Deprecated
 
 - `Modifier::getIdnUriString` use `Modifier::displayUriString` instead
+- `Modifier::getUri` use `Modifier::uri` instead
 
 ### Removed